+++ modulo3/projeto-labex/src/App.js
@@ -1,15 +1,13 @@
 
 import React from 'react'
 import Router from './router/Router';
-import Menu from './components/Menu'
-import Footer from'./components/Footer'
 
 
 
 function App() {
 
   return (
-  
+   
      <Router>
        <App/>
     </Router>
